refactor: move메서드 baseTurn에서 override하는 식으로 변경(공통 코드 제거)
