+++ src/main/java/janggi/domain/turn/ChoTurn.java
@@ -1,9 +1,7 @@
 package janggi.domain.turn;
 
-import janggi.domain.Position;
 import janggi.domain.Side;
 import janggi.domain.board.Board;
-import janggi.domain.piece.PieceAttribute;
 
 public class ChoTurn extends BaseTurn {
     public ChoTurn(Board board, int turn) {
@@ -11,22 +9,12 @@ public class ChoTurn extends BaseTurn {
     }
 
     @Override
-    public TurnState move(Position start, Position end) {
-        PieceAttribute pieceAttribute = board.move(start, end, Side.CHO);
-
-        if(turn == MAX_TURN) {
-            return new TurnState(new FinishTurn(board, turn + 1, Side.EMPTY), pieceAttribute);
-        }
-
-        if (board.isEndGame()) {
-            return new TurnState(new FinishTurn(board, turn + 1, Side.HAN), pieceAttribute);
-        }
-
-        return new TurnState(new HanTurn(board, turn + 1), pieceAttribute);
+    public Side getCurrentSide() {
+        return Side.CHO;
     }
 
     @Override
-    public Side getCurrentSide() {
-        return Side.CHO;
+    protected PlayerTurn nextTurn() {
+        return new HanTurn(board, turn + 1);
     }
 }
